AT1-464 Install all TMC packages in Dockerfile
In order to use the console scripts created during installation,
the packages need to be installed.  These console scripts are
the used to run the Tango device servers.

If we don't install the packages, we can't use relative imports.
E.g., `csp_subarray_leaf_node.py` cannot be
run directly without installing.

Also, the packages need to be installed to run `pytest` directly,
as done for the unit test job.