AT1-464 Fix CspSubarrayLeafNode launch script
Changed server name from `csp_subarray_leaf_node` to
`CspSubarrayLeafNodeDS` in dsconfig JSON file to match the new
executable name.  The "DS" suffix stands for Dish Server.  This is
to avoid confusion with the device class name `CspSubarrayLeafNode`.

The `py_modules` line was removed from `setup.py` because the modules
are no longer directly in the `src` folder.