Fix dish simulator achievedPointing timestamps
When setting the tango-simlib model quantity, the timestamp
must be in seconds since the epoch, not milliseconds.  If we use
milliseconds, there is numeric overflow and the timestamps for
the published attribute reading events end up in the past!

The value inside the first index of the spectrum attribute is
also a timestamp, but it must be in milliseconds, according to the
TM-DSH ICD.

Also remove logging of the achieved pointing - these values change
very frequently and are better viewed via attribute history archive.